allow RF_EVPN paths
We want to utilize gobgp for type 5 advertisements. We found that
while gobgpd was able to generate the announcement just fine, we were
missing a fundamental bit of functionality to announce all locally
known VNIs and v4 unicast traffic in the EVPN address family.
Essentially this is the equivalent of the functionality provided in
FRR provided by the advertise-all-vni and advertise ipv4 unicast
statements in the EVPN address family.